+++ Assets/Scripts/Combat/DamageCalculator.cs
@@ -9,6 +9,14 @@ namespace HerOClock.Combat
     /// It is a pure function: numbers in, numbers out, and the only state it touches is the
     /// advance of the random sequence. That allows the whole thing to be tested outside Unity
     /// against the examples written in the spec.
+    ///
+    /// **Every step is computed in double, deliberately.** The inputs arrive as floats because
+    /// that is what the sheets hold, but the arithmetic must not stay there. C# allows float
+    /// operations to be carried out at a higher precision than float, and whether an intermediate
+    /// gets rounded back is up to the runtime. That is enough to change a result: 40/100 lands
+    /// either side of 0.4, which turned one attack of exactly 262.5 into 262 outside Unity and
+    /// 263 inside it. Widening to double first removes the ambiguity, which matters because a
+    /// battle is supposed to replay identically anywhere.
     /// </summary>
     public static class DamageCalculator
     {
@@ -40,31 +48,31 @@ namespace HerOClock.Combat
             }
 
             // Step 3: mitigation.
-            float mitigation = Mitigation(input);
+            double mitigation = Mitigation(input);
 
             // Once resistance goes past 100%, the attack turns into healing. In that case there
             // is no damage, so no evasion is rolled, no thorns trigger and no life is stolen.
-            if (mitigation > 100f)
+            if (mitigation > 100.0)
             {
-                result.Healing = Round(input.BaseDamage * (mitigation - 100f) / 100.0);
+                result.Healing = Round(input.BaseDamage * (mitigation - 100.0) / 100.0);
                 result.Thorns = 0;
                 return result;
             }
 
             // Step 4: evasion, rolled exactly once per attack.
-            float evasionFactor = 1f;
+            double evasionFactor = 1.0;
 
             if (random != null && random.Roll(input.TargetEvasionChance))
             {
                 result.Evaded = true;
                 result.PerfectEvasion = random.Roll(PerfectEvasionShare);
 
-                float reduction = result.PerfectEvasion ? PerfectEvasionReduction : NormalEvasionReduction;
-                evasionFactor = 1f - reduction / 100f;
+                double reduction = result.PerfectEvasion ? PerfectEvasionReduction : NormalEvasionReduction;
+                evasionFactor = 1.0 - reduction / 100.0;
             }
 
             // Step 5: final damage. Reductions always multiply together, they never add up.
-            double damage = input.BaseDamage * (1f - mitigation / 100f) * evasionFactor;
+            double damage = input.BaseDamage * (1.0 - mitigation / 100.0) * evasionFactor;
             result.Damage = Math.Max(0, Round(damage));
 
             // Step 6: life steal, on the final damage and only for physical attacks.
@@ -81,18 +89,18 @@ namespace HerOClock.Combat
         /// Physical stays between 0 and the cap. Elemental can go past 100, turning the attack
         /// into healing, or down to -100 through stacked debuffs.
         /// </summary>
-        private static float Mitigation(DamageInput input)
+        private static double Mitigation(DamageInput input)
         {
-            float constant = 50f * Math.Max(1, input.AttackerLevel);
-            float fromCurve = CharacterStats.DiminishingReturns(MitigationCap, input.TargetMitigationPoints, constant);
+            double constant = 50.0 * Math.Max(1, input.AttackerLevel);
+            double fromCurve = CharacterStats.DiminishingReturns(MitigationCap, input.TargetMitigationPoints, constant);
 
             if (input.Type == DamageType.Physical)
             {
                 return fromCurve;
             }
 
-            float total = fromCurve + input.TargetResistanceBonus;
-            return total < -100f ? -100f : total;
+            double total = fromCurve + input.TargetResistanceBonus;
+            return total < -100.0 ? -100.0 : total;
         }
 
         /// <summary>
